Pyrite interfacial energies are now anisotropic with a 4-fold symmetry.
sigma_pyr_??? and sigma_???_pyr values were subscripted with a 0 to represent the global value of surface energy in the shared module. Local interfacial energies (same variables, without the subscript) are now calculated inside pfsolve based on the local interfacial orientation (from delypyr and delzpyr) and theta values.

In a future commit, I will try to do this for all phases, preferably in another subroutine that gets called by pfsolve.